Move task into a separated folder
