+++ lib/faas/handler.js
@@ -2,7 +2,7 @@
 
 const alert = require('./aws-alert')
 const func = require('./aws-func')
-const task = require('./task')
+const task = require('./task-artillery')
 
 const impl = {
   /**
