+++ lib/faas/task-artillery/index.js
