+++ lib/faas/task-artillery/taskResult.js
