+++ lib/faas/task-artillery/taskValid.js
