+++ lib/index.js
@@ -10,7 +10,7 @@ const url = require('url')
 const yaml = require('js-yaml')
 
 const func = require('./faas/aws-func')
-const task = require('./faas/task')
+const task = require('./faas/task-artillery')
 const npm = require('./npm')
 const Serverless = require('./serverless-fx')
 
@@ -21,14 +21,15 @@ const constants = {
   ServerlessDirectories: [
     'aws-alert',
     'aws-func',
+    'task-artillery',
   ],
   ServerlessFiles: [
     'aws-alert/index.js',
     'handler.js',
-    'aws-func/index.js', 'aws-func/define.js', 'aws-func/invoke.js', 'aws-func/handle.js', 'aws-func/valid.js',
+    ...['index.js', 'define.js', 'invoke.js', 'handle.js', 'valid.js'].map(file => path.join('aws-func', file)),
     'package.json', 'package-lock.json',
     'serverless.yml',
-    'task.js', 'taskDef.js', 'taskExec.js', 'taskPlan.js', 'taskResult.js', 'taskValid.js',
+    ...['index.js', 'taskDef.js', 'taskExec.js', 'taskPlan.js', 'taskResult.js', 'taskValid.js'].map(file => path.join('task-artillery', file)),
   ],
   TestFunctionName: 'loadGenerator',
   ScheduleName: '${self:service}-${opt:stage, self:provider.stage}-monitoring', // eslint-disable-line no-template-curly-in-string
