+++ tests/lib/faas/aws-alert/alert.spec.js
@@ -12,7 +12,7 @@ const { expect } = chai
 // eslint-disable-next-line import/no-dynamic-require
 const alert = require(path.join('..', '..', '..', '..', 'lib', 'faas', 'aws-alert'))
 // eslint-disable-next-line import/no-dynamic-require
-const task = require(path.join('..', '..', '..', '..', 'lib', 'faas', 'task.js'))
+const task = require(path.join('..', '..', '..', '..', 'lib', 'faas', 'task-artillery'))
 
 let analysis
 let result
