+++ tests/lib/faas/task-artillery/taskDef.spec.js
@@ -4,7 +4,7 @@ const path = require('path')
 const { expect } = chai
 
 // eslint-disable-next-line import/no-dynamic-require
-const task = require(path.join('..', '..', '..', 'lib', 'faas', 'task.js'))
+const task = require(path.join('..', '..', '..', '..', 'lib', 'faas', 'task-artillery'))
 
 let script
 let expected
