+++ tests/lib/faas/task-artillery/taskPlan.spec.js
@@ -10,9 +10,9 @@ chai.use(sinonChai)
 const { expect } = chai
 
 // eslint-disable-next-line import/no-dynamic-require
-const func = require(path.join('..', '..', '..', 'lib', 'faas', 'aws-func'))
+const func = require(path.join('..', '..', '..', '..', 'lib', 'faas', 'aws-func'))
 // eslint-disable-next-line import/no-dynamic-require
-const task = require(path.join('..', '..', '..', 'lib', 'faas', 'task.js'))
+const task = require(path.join('..', '..', '..', '..', 'lib', 'faas', 'task-artillery'))
 
 let script
 let phase
