+++ tests/lib/faas/task-artillery/taskValid.spec.js
@@ -4,9 +4,9 @@ const path = require('path')
 const { expect } = chai
 
 // eslint-disable-next-line import/no-dynamic-require
-const func = require(path.join('..', '..', '..', 'lib', 'faas', 'aws-func'))
+const func = require(path.join('..', '..', '..', '..', 'lib', 'faas', 'aws-func'))
 // eslint-disable-next-line import/no-dynamic-require
-const task = require(path.join('..', '..', '..', 'lib', 'faas', 'task.js'))
+const task = require(path.join('..', '..', '..', '..', 'lib', 'faas', 'task-artillery'))
 
 const getDefaultSettings = () => {
   const ret = func.define.getSettings()
